Extract message viewer into component
